fix(*): removing some duplicate code
