+++ Notification Agent Alerts/AppDelegate.swift
@@ -27,6 +27,10 @@ class AppDelegate: NSObject, NSApplicationDelegate {
         isConfigured = true
         NSApplication.shared.activate(ignoringOtherApps: true)
         notificationDispatch.startObservingForNotifications()
+        guard !UserNotificationController.shared.agentTriggeredByNotificationCenter else {
+            completion()
+            return
+        }
         efclController.parseArguments()
         completion()
     }
