+++ Shared/Extensions/EFCLController-Extension.swift
@@ -11,47 +11,15 @@ import Foundation
 
 extension EFCLController {
     
-    // MARK: - Internal methods
-    
-    /// Exit the app with the related reason code.
-    /// - Parameter reason: reason why the application should exit.
-    /// - Returns: never.
-    internal func applicationExit(withReason reason: ExitReason) {
-        switch reason {
-        case .untrackedSuccess:
-            exit(200)
-        case .mainButtonClicked, .userFinishedOnboarding:
-            exit(0)
-        case .secondaryButtonClicked:
-            exit(2)
-        case .tertiaryButtonClicked:
-            exit(3)
-        case .userDismissedNotification, .userDismissedOnboarding:
-            exit(239)
-        case .invalidArgumentsSyntax:
-            exit(250)
-        case .invalidArgumentFormat:
-            exit(255)
-        case .internalError, .cancelPressed:
-            exit(1)
-        case .receivedSigInt:
-            exit(201)
-        case .unableToLoadResources:
-            exit(260)
-        case .timeout:
-            exit(4)
-        }
-    }
-    
     // MARK: - Methods
     
     /// Check if the app is running testes for other workflows, debug or different workflow and if not start parsing the launch arguments.
     func parseArguments(_ arguments: [String] = CommandLine.arguments) {
+        guard !arguments.contains("--isRunningTest") else { return }
         do {
-            guard !arguments.contains("--isRunningTest") else { return }
             guard let base64EncodedData = arguments.last,
                   let data = Data(base64Encoded: base64EncodedData) else {
-                applicationExit(withReason: .internalError)
+                Utils.applicationExit(withReason: .internalError)
                 return
             }
             let taskObject = try JSONDecoder().decode(TaskObject.self, from: data)
@@ -62,11 +30,11 @@ extension EFCLController {
         } catch let error {
             guard let efclError = error as? NAError else {
                 logger.log("No recognized error: %{public}@.", [error.localizedDescription])
-                applicationExit(withReason: .internalError)
+                Utils.applicationExit(withReason: .internalError)
                 return
             }
             logger.log("%{public}@", [efclError.localizedDescription])
-            applicationExit(withReason: efclError.efclExitReason)
+            Utils.applicationExit(withReason: efclError.efclExitReason)
         }
     }
 }
